remove 2.10 workarounds from main code
