+++ shared/src/main/scala/org/scalamock/AbstractMockFactoryBase.scala
@@ -1,0 +1,37 @@
+// Copyright (c) 2011-2015 ScalaMock Contributors (https://github.com/paulbutcher/ScalaMock/graphs/contributors)
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package org.scalamock
+
+import org.scalamock.clazz.Mock
+import org.scalamock.context.MockContext
+import org.scalamock.function.MockFunctions
+import org.scalamock.matchers.Matchers
+
+/** ScalaMock public interface */
+trait AbstractMockFactoryBase extends Mock with MockFunctions with Matchers { this: MockContext =>
+
+  protected def withExpectations[T](what: => T): T
+
+  protected def inAnyOrder[T](what: => T): T
+  protected def inSequence[T](what: => T): T
+  protected def inAnyOrderWithLogging[T](what: => T): T
+  protected def inSequenceWithLogging[T](what: => T): T
+}
