+++ shared/src/main/scala/org/scalamock/util/MacroAdapter.scala
